Managing bpf centrally configuration through salt
Salt states to centrally manage the bpf configuration files. Will not
handle restarting the NSM services when the bpf configuration changes.

This requires the addition of a grain for each minion functioning as
a sensor, or else the run will fail on the minion. Just add a list of
the interfaces that are monitoring to a grain named "sensor_interfaces".
This can be done by modifying the /etc/salt/grains file like so -

sensor_interfaces:
  - <monitoring interface_1>
  - <monitoring interface_2>

If you can think of a different way to tell salt how many
sensor_interfaces a minion has let me know. Custom grain seemed to be
the most straight forward solution.

Modify the salt/sensor/init.sls file to enable the bpf configuration.

Explanation of folder layout for bpf management is straight forward and
included in the sls file.