Improvements to the bpf config interface selection
Using the 'ip_interfaces' grain to get the list of network interfaces on
each minion, filtering out the loopback interface with a conditional
check. We still will check and see if the 'sensor_interface' grain has
been configured and use that if it has.

Updated the comments associated with the file to handle a potential edge
case on sensors where sosetup was run before roughly July 22nd. This was
when Security Onion (I believe) started configuring all interfaces and
then enabling the ones users selected through the sosetup. [1]

[1] - http://blog.securityonion.net/2014/07/new-securityonion-setup-package.html